fix: Ensure subscriber query works with ruby keyword args
When upgrading codebase-website to Ruby 3.2 I discovered the blog newsletter subscribe functionality didn't work.

Turns out calling this query to subscribe was generating the error:
Wrong number of args got 2 expected 1

All the other queries generated in this abe client lib work fine, the difference
is the subscriber wrapped it's args in { }. The underlying graphql-client gem
expects variables as a keyword arg. By removing the { } it fixes the issue.